+++ src/app/user-card/user-card.component.ts
@@ -1,6 +1,7 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { User } from './user-card.model';
 import { UserService } from 'src/services/user.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-user-card',
@@ -13,12 +14,28 @@ export class UserCardComponent {
   constructor(private userService: UserService) {}
 
   handleDeleteUser(id: number) {
-    console.log(id);
-    this.userService.deleteUser(id).subscribe({
-      next: (value) => {
-        console.log(value);
-        this.deletedUserEvent.next(id);
-      },
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.userService.deleteUser(id).subscribe({
+          next: () => {
+            Swal.fire({
+              title: 'Deleted!',
+              text: 'Your file has been deleted.',
+              icon: 'success',
+            }).then(() => {
+              this.deletedUserEvent.next(id);
+            });
+          },
+        });
+      }
     });
   }
